Spawn Manager updated
Spawn manager now spawns the tank in the proper place, however it is
facing a weird direction (not that that matters). Still need to code
the spawning of enemies and allies. The problem right now is that the
tank prefab has a camera attached to it, and the allies and enemies
can’t have that main camera.

Also, the tank prefab doesn’t have a controller script attached to it,
and it really should.